More renames and moves
