+++ package.json
@@ -12,14 +12,15 @@
   },
   "dependencies": {},
   "devDependencies": {
-    "typescript": "^4.3.2",
-    "vite": "^2.5.1",
     "@babel/core": "^7.9.6",
     "@babel/preset-env": "^7.9.6",
+    "@babel/preset-typescript": "^7.15.0",
     "babel-jest": "^26.0.1",
     "eslint": "^7.0.0",
     "eslint-config-airbnb-base": "^14.1.0",
     "eslint-plugin-import": "^2.20.2",
-    "jest": "^26.0.1"
+    "jest": "^26.0.1",
+    "typescript": "^4.3.2",
+    "vite": "^2.5.1"
   }
 }
