+++ src/app/game.consts.ts
